feat: Divide to CollectionView and Stock NameSpace from NameSpace enum type
