+++ JuiceMaker/JuiceMaker/NameSpace/StockNameSpace.swift
@@ -1,0 +1,18 @@
+//
+//  StockNameSpace.swift
+//  JuiceMaker
+//
+//  Created by Oh Donggeon on 2022/02/19.
+//
+
+import Foundation
+
+enum StockNameSpace {
+    static var initValue: [Fruit:Int] {
+        [.strawberry: 10,
+         .pineapple: 10,
+         .mango: 10,
+         .kiwi: 10,
+         .banana: 10]
+    }
+}
